// src/frontend/src/features/Dashboard/features/Main/components/RightSidebar/types.ts
export interface BasicInfoData {
    name: string;
    description?: string;
    icon?: string;
}

export interface CustomizationData {
    iconColor?: string;
    cardColor?: string;
    navbarColor?: string;
    backgroundColor?: string;
    leftSidebarColor?: string;
    rightSidebarColor?: string;
    textColor?: string;
}